restructure to flatten classes; no "new" for modules
customize to include hostname for msg and date for stream; batch send data to cw
