+++ Gemfile
@@ -1,8 +1,3 @@
 source "https://rubygems.org"
 
 gemspec
-
-platforms :jruby do
-  gem 'jruby-openssl'
-  gem 'multi_json', '~> 1' # https://github.com/bundler/bundler/issues/4157
-end
