+++ lib/cloudwatchlogger.rb
@@ -1,20 +1,14 @@
-require File.join(File.dirname(__FILE__), 'cloudwatchlogger', 'client')
-
+require 'cloudwatchlogger/client'
 require 'logger'
 
 module CloudWatchLogger
   class LogGroupNameRequired < ArgumentError; end
   class LogEventRejected < ArgumentError; end
 
-  def self.new(credentials, log_group_name, log_stream_name = nil, opts = {})
+  def self.setup_logger(log_group_name, log_stream_name = nil, credentials = {}, opts = {})
     client = CloudWatchLogger::Client.new(credentials, log_group_name, log_stream_name, opts)
     logger = Logger.new(client)
-
-    if client.respond_to?(:formatter)
-      logger.formatter = client.formatter
-    elsif client.respond_to?(:datetime_format)
-      logger.datetime_format = client.datetime_format
-    end
+    logger.formatter = client.formatter
 
     logger
   end
