+++ lib/cloudwatchlogger/version.rb
@@ -1,3 +1,3 @@
 module CloudWatchLogger
-  VERSION = '0.2.2'.freeze
+  VERSION = '1.0.0'
 end
